LPS-55753 Fix Tests
